commands: move show_predecessor_patch code to obslog.rs
